batch_inference/requirements.txt
apache-airflow~=2.5.3
pandas~=1.5.3
scikit-learn~=1.2.1
pydantic~=1.10.4
minio==7.1.14
